requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0
requests>=2.24.0,<2.25.0